Add reason argument to set_status() so that custom messages flow
Without adding a named argument of `reason` to tornado's `set_status()`
method, messages associated with a thrown HTTPError will not be returned
to the caller.  Instead, the reason corresponding to the built-in status
code is returned.  In addition, because the reason argument is used,
enforcement of a built in status code is skipped, thereby allowing for
the optional use of non-builtin status codes.

Updated the TestableJSONErrorsHandler class to have a compatible signature
to Torado's `set_status()` method (to include the `reason` parameter).

Fixes #283