Fix bug when other jupyter server running
It is possible to receive ZMQ messages from the kernel before the kernel
gateway is ready to receive them. This leads to a KeyError in pool.py

To mitigate this we check if the kernel id is in the list of available
callbacks and return early if not.

Fixes #277 and fixes #288